Put layer pattern and bitrate config in JSON.
This is important because a single-temporal-layer encoding can predict
more efficiently than the base layer of a 3-tl encoding, since the
single-layer encoding can predict from all buffers but tl0 from a 3-tl
encoding can only predict from last.

Also storing bitrate config because the performance of the top
layer of 800:1500:2500 kbps isn't the same as 10:100:2500 kbps, so the
bitrate used for the layer itself isn't enough to fairly characterize
it.