Add hashing of input file + scrub input path.
Adds hashing of the input file with sha1sum to prevent version
mismatches of files without knowing it.

Also adds scrubbing of input file path (e.g. prints basename of the
input file only). The path of a file is likely not very informative when
sharing results externally and potentially leaks data.