Add --dump-commands to generate_data.py.
Outputs all encoder commands instead of running them, which is useful
when reproducing and fixing issues in encoders.

Also fixes issue found when testing where running an unexisting command
(because of missing dependencies, etc.) would throw an exception and
kill the worker thread instead of instead of generating an error with
the command in the string.